% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activateRecord.R
\name{activateRecord}
\alias{activateRecord}
\title{Activate or Deactivate Report Users}
\usage{
activateRecord(
  oid,
  active,
  event_user,
  table_name,
  event_table_name,
  parent_field_name
)
}
\arguments{
\item{oid}{\code{integerish(1)}. The OID of the record to modify.}

\item{active}{\code{logical(1)}. The value to set in the database.}

\item{event_user}{\code{integerish(1)}. The User OID of the user performing the action.}

\item{table_name}{\code{character(1)}. The name of the table for which the
object is being activated/deactivated.}

\item{event_table_name}{\code{character(1)}. The name of the table in which
events are stored for the table named in \code{table_name}.}

\item{parent_field_name}{\code{character(1)}. The name of the foreign key
field mapping between \code{event_table_name} and \code{table_name}.}
}
\description{
Changes the value of a table's IsActive property.
}
\details{
When the value of \code{active} matches what is already in the database,
no action will be taken. Otherwise, the value is updated and an event
is recorded.
}
