% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/activateRecord.R
\name{activateReportUser}
\alias{activateReportUser}
\alias{activateRecord}
\title{Activate or Deactivate Report Users}
\usage{
activateRecord(
  oid,
  active,
  event_user,
  table_name,
  event_table_name,
  parent_field_name
)
}
\arguments{
\item{oid}{\code{integerish(1)}. The OID of the ReportUser to modify.}

\item{active}{\code{logical(1)}. The value to set in the database.}

\item{event_user}{\code{integerish(1)}. The OID of the ReportUser performing the action.}
}
\description{
Changes the value of the ReportUser.IsActive property.
}
\details{
When the value of \code{active} matches what is already in the database,
no action will be taken. Otherwise, the value is updated and an event
is recorded.
}
