% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditApplicationSetting.R
\name{addEditApplicationSetting}
\alias{addEditApplicationSetting}
\title{Add or Edit Application Settings}
\usage{
addEditApplicationSetting(
  oid = numeric(0),
  setting_key,
  setting_value,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of an application setting. If
length is zero, a new setting is added.}

\item{setting_key}{\code{character(1)}. The name of the setting key. Whitespace
will be trimmed off the ends.}

\item{setting_value}{\code{character(1)}. The value for the setting. Whitespace
will be trimmed off the end.}

\item{event_user}{\code{integerish(1)}. The OID of the user adding or editing
the setting.}
}
\description{
Provides the user with tools to add or edit application
setting objects.
}
