% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditDateReportingFormat.R
\name{addEditDateReportingFormat}
\alias{addEditDateReportingFormat}
\alias{addEditSchedule}
\title{Add or Edit a DateReportingFormat Object}
\usage{
addEditSchedule(
  oid = numeric(0),
  schedule_name,
  frequency,
  frequency_unit,
  offset_overlap,
  offset_overlap_unit,
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the DateReportingFormat to be edited.
When the length is 0, a new DateReportingFormat will be added.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the record will be marked as
active.}

\item{event_user}{\code{integerish(1)}. the User.OID of the user performing
the action.}

\item{format_name}{\code{character(1)}. The name of the reporting format.}

\item{description}{\code{character(1)}. An option description of the format. Its
recommended use is to give an example of what the format will look like
when rendered in a report.}

\item{format_code}{\code{character(1)}. The code used to format the date/time.
Use the formats described in \code{\link[=strptime]{strptime()}}.}

\item{increment_start}{\code{integerish(1)}. An optional amount of time to
increment the starting time when reported. In almost all cases, 0 will be
suitable.}

\item{increment_start_unit}{\code{character(0)}. The unit of time to apply to
\code{increment_start}. One of \code{c("Second", "Minute", "Hour", "Day", "Week", "Month", "Year")}.}

\item{increment_end}{\code{integerish(1)}. An optional amount of time to
increment the ending time when report. See Details.}

\item{increment_end_unit}{\code{character(0)}. The unit of time to apply to
\code{increment_end}.  One of \code{c("Second", "Minute", "Hour", "Day", "Week", "Month", "Year")}.}
}
\description{
Adds a new DateReportingFormat object or edits an existing
DateReportingFormat objects in the database.
}
