% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditFileArchive.R
\name{addEditFileArchive}
\alias{addEditFileArchive}
\alias{addFileArchive}
\alias{editFileArchive}
\title{Add or Edit Files to the FileArchive}
\usage{
addFileArchive(
  parent_report_template = NA_real_,
  parent_report_instance = NA_real_,
  description = "",
  is_logo = FALSE,
  file_path,
  file_name = tools::file_path_sans_ext(basename(file_path)),
  file_extension = tools::file_ext(file_name)
)

editFileArchive(
  oid,
  parent_report_template = numeric(0),
  parent_report_instance = numeric(0),
  description = character(0),
  file_name = character(0),
  is_logo = logical(0)
)
}
\arguments{
\item{parent_report_template}{\code{integerish(0/1)}. The ReportTemplate OID
associated with the file. Use \code{NA} for an unassociated file. Must
be length 1 for \code{addFileArchive}. For \code{editFileArchive}, a length
1 object is used prompt an update.}

\item{parent_report_instance}{\code{integerish(0/1)}. The ReportInstance OID
associated with the file. Use \code{NA} for an unassociated file. Must
be length 1 for \code{addFileArchive}. For \code{editFileArchive}, a length
1 object is used prompt an update.}

\item{description}{\code{character(0/1)}. An optional description to provide
for the file.}

\item{is_logo}{\code{logical(0/1)}. When \code{TRUE}, the file is flagged as a
logo file. Must be length 1 for \code{addFileArchive}. For \code{editFileArchive},
a length 1 object is used prompt an update.}

\item{file_path}{\code{character(1)}. The path to the file being uploaded.
The file will be converted to a \code{raw} vector.}

\item{file_name}{\code{character(0/1)}. The name of the file to store in the
database. If length 0, it will be extracted from \code{file_path}. Must
be length 1 for \code{addFileArchive}. For \code{editFileArchive}, a length
1 object is used prompt an update.}

\item{file_extension}{\code{character(1)}. The file extension. If length 0,
it will be extracted from \code{file_path}.}

\item{oid}{\code{integerish(0/1)}. The OID of the FileArchive object to edit.
When this has length 0, a new file is uploaded.}
}
\description{
Add or Edit Files to the FileArchive
}
\details{
Enables the user to upload a file to the File Archive.
Also used to edit the file name or description of an existing
FileArchive object. Edits to FileArchive objects are
not event tracked.
}
