% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditLogo.R
\name{addEditLogo}
\alias{addEditLogo}
\alias{addLogo}
\alias{editLogo}
\title{Add a Logo File or Edit Properties of an Existing Logo}
\usage{
addLogo(
  file_path,
  description = "",
  file_name = tools::file_path_sans_ext(basename(file_path)),
  file_extension = tools::file_ext(file_name)
)

editLogo(oid, description = character(0), file_name = character(0))
}
\arguments{
\item{file_path}{\code{character(1)}. The path to the file being uploaded.
The file will be converted to a \code{raw} vector.}

\item{description}{\code{character(1)}. An optional description to provide
for the file.}

\item{file_name}{\code{character(1)}. The name of the file to store in the
database.}

\item{file_extension}{\code{character(1)}. The file extension of the file.
This will typically be drawn from the file being uploaded.}

\item{oid}{\code{integerish(1)}. The OID of the FileArchive object to edit.}
}
\description{
Add a Logo File or Edit Properties of an Existing Logo
}
\details{
Enables the user to upload a file to the File Archive to use
as a logo in report titles. Also used to edit the file name or
description of an existing logo file. Edits to logo files are
not event tracked.
}
