% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportInstance.R
\name{addEditReportInstance}
\alias{addEditReportInstance}
\title{Add and Edit ReportInstance Records}
\usage{
addEditReportInstance(
  report_instance_oid = numeric(0),
  parent_report_template = numeric(0),
  start_time,
  end_time,
  is_signature_required,
  is_scheduled,
  instance_title,
  is_submitted,
  event_user
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(0/1)}. The OID of the report
instance to be edited. When length 0, a new record is added.}

\item{parent_report_template}{\code{integerish(0/1)}. The OID of the parent
report template. This will be ignored with the instance OID is known,
but is required when adding a new instance.}

\item{start_time}{\code{POSIXct(1)}. The start time for the reporting period.}

\item{end_time}{\code{POSIXct(1)}. The end time for the reporting period.}

\item{is_signature_required}{\code{logical(1)}. Indicates if signatures must
be given prior to submission.}

\item{is_scheduled}{\code{logical(1)}. Indicates if the instance is a scheduled
instance by the template's schedule settings.}

\item{instance_title}{\code{character(1)}. The title of the instance.}

\item{is_submitted}{\code{character(1)}. Indicates if the instance has been
submitted.}
}
\description{
Provides the user with the tools to add new report instances
or edit existing instances.
}
