% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportInstanceDistribution.R
\name{addEditReportInstanceDistribution}
\alias{addEditReportInstanceDistribution}
\title{Add or Edit ReportInstanceDistribution Objects}
\usage{
addEditReportInstanceDistribution(
  oid = numeric(0),
  parent_report_instance,
  parent_user = numeric(0),
  parent_role = numeric(0),
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReportInstanceDistribution
object to be edited. Use \code{numeric(0)} to add a new object.}

\item{parent_report_instance}{\code{integerish(1)}. The OID of the ReportInstance
object being associated with a disclaimer.}

\item{parent_user}{\code{integerish(0/1)}. The OID of the User object being
designated as a recipient. Only one of \code{parent_user} or \code{parent_role}
may be given.}

\item{parent_role}{\code{integerish(0/1)}. The OID of the Role object being
designated as a recipient. Only one of \code{parent_user} or \code{parent_role}
may be given.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the association will be marked
as active.}

\item{event_user}{\code{integerish(1)}. The OID of the User performing the action.}

\item{order}{\code{integerish(1)}. The ordered position in which the
signature is displayed on the report.}
}
\description{
Enables the user to add ReportInstanceDistribution objects
or edit an existing ReportInstanceDistribution object. Also manages the
recording of events associated with the action.
}
