% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportInstanceNarrative.R
\name{addEditReportInstanceNarrative}
\alias{addEditReportInstanceNarrative}
\title{Add and Edit ReportInstanceNarrative Records}
\usage{
addEditReportInstanceNarrative(report_instance_oid, narrative, event_user)
}
\arguments{
\item{report_instance_oid}{\code{integerish(0/1)}. The OID of the report
instance to be for which the narrative is to be edited.}

\item{narrative}{\code{character(1)}. The text of the narrative. Whitespace
will be trimmed before sending to the database. If, after trimming,
the text is an empty string, it will still be sent to the database as
that may be a valid edit.}

\item{event_user}{\code{integerish(1)}. The OID of the user making changes to
the narrative.}
}
\description{
Provides the user with the tools to add or edit the narrative
associated with a report instance.
}
\details{
All changes to the narrative are recorded as an edit. The
ReportInstanceNarrativeEvent table records any save events, even if no
actual changes are made.
}
