% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportTemplateDisclaimer.R
\name{addEditReportTemplateDisclaimer}
\alias{addEditReportTemplateDisclaimer}
\title{Add or Edit ReportTemplateDisclaimer Objects}
\usage{
addEditReportTemplateDisclaimer(
  oid = numeric(0),
  parent_report_template,
  parent_disclaimer,
  order,
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReportTemplateDisclaimer
object to be edited. Use \code{numeric(0)} to add a new object.}

\item{parent_report_template}{\code{integerish(1)}. The OID of the ReportTemplate
object being associated with a disclaimer.}

\item{parent_disclaimer}{\code{integerish(1)}. The OID of the Disclaimer
object being associated with the ReportTemplate.}

\item{order}{\code{integerish(1)}. The ordered position in which the
disclaimer is displayed on the report.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the association will be marked
as active.}

\item{event_user}{\code{integerish(1)}. The OID of the User performing the action.}
}
\description{
Enables the user to add ReportTemplateDisclaimer objects
or edit an existing ReportTemplateDisclaimer object. Also manages the
recording of events associated with the action.
}
