% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportTemplateFooter.R
\name{addEditReportTemplateFooter}
\alias{addEditReportTemplateFooter}
\title{Add or Edit ReportTemplateFooter Objects}
\usage{
addEditReportTemplateFooter(
  oid = numeric(0),
  parent_report_template,
  parent_footer,
  order,
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReportTemplateFooter
object to be edited. Use \code{numeric(0)} to add a new object.}

\item{parent_report_template}{\code{integerish(1)}. The OID of the ReportTemplate
object being associated with a disclaimer.}

\item{parent_footer}{\code{integerish(1)}. The OID of the Footer
object being associated with the ReportTemplate.}

\item{order}{\code{integerish(1)}. The ordered position in which the
footer is displayed on the report.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the association will be marked
as active.}

\item{event_user}{\code{integerish(1)}. The OID of the User performing the action.}
}
\description{
Enables the user to add ReportTemplateFooter objects
or edit an existing ReportTemplateFooter object. Also manages the
recording of events associated with the action.
}
