% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportTemplatePermission.R
\name{addEditReportTemplatePermission}
\alias{addEditReportTemplatePermission}
\title{Add or Edit ReportTemplatePermission Objects}
\usage{
addEditReportTemplatePermission(
  oid = numeric(0),
  parent_report_template = numeric(0),
  parent_role = numeric(0),
  can_view,
  can_add_notes,
  can_edit_narrative,
  can_submit,
  can_start_revision,
  is_active,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReportTemplateDisclaimer
object to be edited. Use \code{numeric(0)} to add a new object.}

\item{parent_report_template}{\code{integerish(0/1)}. The OID of the ReportTemplate
object being associated with a permission.}

\item{parent_role}{\code{integerish(0/1)}. The OID of the Permission
object being associated with the ReportTemplate.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the association will be marked
as active.}

\item{event_user}{\code{integerish(1)}. The OID of the User performing the action.}
}
\description{
Enables the user to add ReportTemplatePermission objects
or edit an existing ReportTemplateDisclaimer object. Also manages the
recording of events associated with the action.
}
