% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportTemplateSignature.R
\name{addEditReportTemplateSignature}
\alias{addEditReportTemplateSignature}
\alias{addEditReportTemplateFooter}
\title{Add or Edit ReportTemplateSignature Objects}
\usage{
addEditReportTemplateFooter(
  oid = numeric(0),
  parent_report_template,
  parent_role,
  order,
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReportTemplateSignature
object to be edited. Use \code{numeric(0)} to add a new object.}

\item{parent_report_template}{\code{integerish(1)}. The OID of the ReportTemplate
object being associated with a disclaimer.}

\item{parent_role}{\code{integerish(1)}. The OID of the Role object being
designated as a required signatory.}

\item{order}{\code{integerish(1)}. The ordered position in which the
signature is displayed on the report.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the association will be marked
as active.}

\item{event_user}{\code{integerish(1)}. The OID of the User performing the action.}
}
\description{
Enables the user to add ReportTemplateSignature objects
or edit an existing ReportTemplateSignature object. Also manages the
recording of events associated with the action.
}
