% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportUser.R
\name{addEditReportUser}
\alias{addEditReportUser}
\title{Add or Edit a ReportUser Entry}
\usage{
addEditReportUser(
  oid = numeric(0),
  last_name,
  first_name,
  login_id,
  email,
  is_internal = FALSE,
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReortUser entry to be
edited. Use \code{numeric(0)} to add a new entry.}

\item{last_name}{\code{character(1)}. The last name of the user.}

\item{first_name}{\code{character(1)}. The first name of the user.}

\item{login_id}{\code{character(1)}. The login ID of the user.}

\item{email}{\code{character(0/1)}. The e-mail address of the user. If not
provided, the user will not receive reports distributed by e-mail but
may use the application otherwise.}

\item{is_internal}{\code{logical(1)}. When \code{TRUE}, the user is considered
internal and may receive incomplete reports. When \code{FALSE}, only reports
that have completed signatures will be distributed.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the user has access to the
application and may receive e-mails with distributed reports. When
\code{FALSE}, access is denied and no e-mails may be sent to the user.}

\item{event_user}{\code{integerish(1)}. The ReportUser.OID of the user performing
the action.}
}
\description{
Executes the SQL statement to add or edit an existing
ReportUser while adding the appropriate events.

When editing, values are compared against what is currently in the
database and only changed values are edited.
}
