% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditRole.R
\name{addEditRole}
\alias{addEditRole}
\title{Add or Edit a Role}
\usage{
addEditRole(
  oid = numeric(0),
  role_name,
  role_description = NA_character_,
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the role to be edited. When
length is 0, a new role will be added.}

\item{role_name}{\code{character(1)}. The name of the role.}

\item{role_description}{\code{character(1)}. The description of the role.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the role will be made
active. Otherwise, it is inactive.}

\item{event_user}{\code{integerish(1)}. The OID of the user performing the}
}
\description{
Adds a new role or edits an existing role in the database.
}
