% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditUserRole.R
\name{addEditUserRole}
\alias{addEditUserRole}
\title{Add and Edit User Role Assignments}
\usage{
addEditUserRole(
  oid = numeric(0),
  parent_user,
  parent_role,
  is_active = TRUE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the UserRole object to modify.
Use \code{numeric(0)} to add a new record.}

\item{parent_user}{\code{integerish(1)}. The OID of the User in the UserRole
assignment.}

\item{parent_role}{\code{integerish(1)}. the OID of the Role in the UserRole
assignment.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the assignment is active.}

\item{event_user}{\code{integerish(1)}. The OID of the User performing the action.}
}
\description{
Add UserRole assignments or edit existing UserRole assignments.
}
