% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReportInstanceGeneration.R
\name{addReportInstanceGeneration}
\alias{addReportInstanceGeneration}
\title{Add Records for Report Instance Previews}
\usage{
addReportInstanceGeneration(
  report_instance_oid,
  report_template_oid,
  start_date_time,
  end_date_time,
  report_format = c("preview", "html", "pdf", "shiny"),
  include_data,
  is_preview,
  is_distributed,
  is_archived,
  is_submission,
  user_oid,
  preview_date_time = Sys.time()
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance
being previewed.}

\item{report_template_oid}{\code{integerish(1)}. The OID of the report template
of which an instance is being previewed.}

\item{start_date_time}{\code{POSIXct(1)}. The start time of the reporting period.}

\item{end_date_time}{\code{POSIXct(1)}. The end time of the reporting period.}

\item{report_format}{\code{character(1)}. The format of the instance being produced by
the application.  One of \code{c("preview", "html", "pdf", "shiny")}.}

\item{include_data}{\code{logical(1)}. Is the data being included with the
download of the preview. Only applies to the HTML and PDF preview
downloads.}

\item{is_preview}{\code{logical(1)}. Indicates if the action is generating a
preview of the report. Previews cannot be archived in the database; may
not be distributed; and is not submitted externally.}

\item{is_distributed}{\code{logical(1)}. Indicates if the instance was distributed
to the users on the distribution list.}

\item{is_archived}{\code{logical(1)}. Indicates if the instance was saved to
the archive.}

\item{is_submission}{\code{logical(1)}. Indicates if the instance was submitted
externally.}

\item{user_oid}{\code{integerish(1)}. The OID of the user initiating the
preview.}

\item{preview_date_time}{\code{POSIXct(1)}. The date and time the preview was
initiated.}
}
\description{
Provides the user with tools to add records for when an
instance is previewed.
}
