% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReportInstanceGenerationRecipient.R
\name{addReportInstanceGenerationRecipient}
\alias{addReportInstanceGenerationRecipient}
\title{Record Recipients to Whom a Report was Sent}
\usage{
addReportInstanceGenerationRecipient(report_instance_generation_oid, user_oid)
}
\arguments{
\item{report_instance_generation_oid}{\code{integerish(1)}. The OID of the report
instance generation.}

\item{user_oid}{\code{integerish}. The OIDs of the users receiving the
report instance.}
}
\description{
Makes a record tying a recipient/user to an attempt to deliver
a report to that user.
}
