% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReportInstanceNote.R
\name{addReportInstanceNote}
\alias{addReportInstanceNote}
\title{Add a Note to a Report Instance}
\usage{
addReportInstanceNote(
  report_instance_oid,
  parent_user,
  note,
  note_date_time = Sys.time()
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance.}

\item{parent_user}{\code{integerish(1)}. The user adding the note.}

\item{note}{\code{character(1)}. The text of the note to be added. If the note
consists of entirely whitespace, the function will silently skip writing
any record to the database. White space will be trimmed before writing
to the database.}

\item{note_date_time}{\code{POSIXct(1)}. The date and time the note is added.}
}
\description{
Adds a note to the collection of notes associated with a
report instance.
}
