% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReportInstanceRevision.R
\name{addReportInstanceRevision}
\alias{addReportInstanceRevision}
\title{Add a Revision Record for a Submitted Report}
\usage{
addReportInstanceRevision(
  report_instance_oid,
  parent_user,
  revision_date_time,
  reason
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance.}

\item{parent_user}{\code{integerish(1)}. The OID of the user initiating the
revision.}

\item{revision_date_time}{\code{POSIXct(1)}. The time stamp the revision is
initiated.}

\item{reason}{\code{character(1)}. The reason for the revision.}
}
\description{
Adds a record to the ReportInstanceRevision table. These
records indicate the start of a revision of a previously submitted report.
}
