% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addReportInstanceSignature.R
\name{addReportInstanceSignature}
\alias{addReportInstanceSignature}
\title{Add a Signature Event to a Report Instance}
\usage{
addReportInstanceSignature(
  report_instance_oid,
  report_template_signature,
  signed,
  event_user,
  signed_datetime = Sys.time()
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance
being signed (or unsigned).}

\item{signed}{\code{logical(1)}. If \code{TRUE}, the report will be marked as signed.
If \code{FALSE}, the signature is reversed.}

\item{event_user}{\code{integerish(1)}. The OID of the user signing the report.}
}
\description{
Provides the user with the tools to add or remove signatures
from a report instance.
}
