% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/castColumn.R
\name{castColumn}
\alias{castColumn}
\title{Convert SQL Columns to R Data Types}
\usage{
castColumn(data, table_name, sql_flavor, conn, ...)
}
\arguments{
\item{data}{\code{data.frame}. The data retuned from SQL.}

\item{table_name}{\code{character(1)}. The name of the table}

\item{sql_flavor}{\code{character(1)}. The type of SQL connection.}

\item{conn}{A database connection object.}
}
\description{
Some SQL drivers do not convert the SQL data types
to R data types. This function performs that work to provide
a consistent result when retrieving data.
}
