% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapsibleDiv.R
\name{collapsibleDiv}
\alias{collapsibleDiv}
\title{Add Elements to a Collapsible Div Tag}
\usage{
collapsibleDiv(
  title,
  id,
  ...,
  style = "",
  checked = FALSE,
  div_id = sprintf("colldiv-\%s", id)
)
}
\arguments{
\item{title}{\code{character(1)} The title for the section.}

\item{id}{\code{character(1)} The id name for the input control that
opens and closes the collapsible.}

\item{...}{UI elements to include in the collapsed portion}

\item{style}{\code{character(1)} Additional CSS styling.}

\item{checked}{\code{logical(1)} Should the div initialize in it's opened
state?}

\item{div_id}{\code{character(1)} The id name for the the collapsible div.}
}
\description{
Creates a UI element with a title bar that can be clicked
to expand or reduce the content from view.
}
\author{
Benjamin Nutter
}
