% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareValues.R
\name{compareValues}
\alias{compareValues}
\title{Compare Values with Consideration for Missing Values}
\usage{
compareValues(current, new)
}
\arguments{
\item{current}{\code{atomic}. Values to be compared against \code{new}.}

\item{new}{\code{atomic}. Values to be compared against \code{current}.}
}
\description{
Determines if there are pair-wise differences between
two vectors while accounting for missing values.
}
