% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/configureReportManager.R
\name{configureReportManager}
\alias{configureReportManager}
\title{Configure the Report Manager}
\usage{
configureReportManager(
  sql_flavor = NULL,
  database_file = NULL,
  driver = NULL,
  server = NULL,
  database = NULL
)
}
\arguments{
\item{database_file}{\code{character(1)}. For SQLite connections, the file
containing the database.}

\item{driver}{\code{character(1)}. For SQL Server connections, the driver to
use when connecting to the database.}

\item{server}{\code{character(1)}. For SQL Server connections, the server
to connect to.}

\item{database}{\code{character(1)}. For SQL Server connections, the database
to connect to.}

\item{flavor}{\code{character(1)}. The flavor of SQL used. This will
determine the SQL engine used. One of \code{c("sql_server", "sqlite")}.}
}
\value{
Invisibly returns \code{NULL}
}
\description{
Assists the user is setting options for the report manager.
This function is called for its side effects in setting options
for the ReportManager application.
}
\details{
For all arguments, \code{NULL} is an acceptable value. When \code{NULL}
is provided, the option associated with that argument will not be altered.
}
