% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{constants}
\alias{constants}
\alias{SUPPORTED_SQL_FLAVOR}
\title{Constants for the ReportManager application}
\format{
An object of class \code{character} of length 2.
}
\usage{
SUPPORTED_SQL_FLAVOR
}
\description{
Commonly used values to simplify maintenance of the package.
}
\keyword{datasets}
