% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataframeToHtml.R
\name{dataframeToHtml}
\alias{dataframeToHtml}
\title{Convert a Data Frame to HTML for Display in Shiny}
\usage{
dataframeToHtml(dframe)
}
\arguments{
\item{dframe}{\code{data.frame}. The data frame to convert to HTML}
}
\description{
Convert a data frame to HTML code
}
