% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dateTimeInput.R
\name{dateTimeInput}
\alias{dateTimeInput}
\alias{useDateTimeInput}
\title{Date/Time Input for Shiny Applications}
\source{
\url{http://www.daterangepicker.com/#example2}
}
\usage{
dateTimeInput(
  inputId,
  label,
  start = NULL,
  end = NULL,
  single_date = TRUE,
  timepicker = TRUE,
  ...
)

useDateTimeInput()
}
\arguments{
\item{inputId}{\code{character(1)} The name for the input object.}

\item{label}{\code{character(1)} The label to display.}

\item{start}{\code{POSIXct(1)} The start date/time for the control. If
\code{NULL}, the current date at midnight is used.}

\item{end}{\code{POSIXct(1)} The end date/time for the control. If
\code{NULL}, the next day at midnight is used.}

\item{single_date}{\code{logical(1)} When \code{FALSE}, two date/times
are input.  When \code{TRUE}, only a single date/time is input.}

\item{timepicker}{\code{logical(1)} Use \code{FALSE} to suppress the
time selection (although, at this point, you may as well use
\code{dateInput}).}

\item{...}{Additional arguments to pass to \code{tags$input}}
}
\description{
Draws a date/time input for a shiny application.
}
\author{
Benjamin Nutter
}
