% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generateReportFiles.R
\name{generateReportFiles}
\alias{generateReportFiles}
\alias{makeReportPreview}
\alias{makeReportForArchive}
\title{Generate a Report and Supplemental Files}
\usage{
makeReportPreview(
  report_instance_oid,
  zipfile,
  include_image = FALSE,
  include_data = FALSE,
  build_dir = tempdir(),
  params = list(),
  report_format = c("html", "pdf")
)

makeReportForArchive(
  report_instance_oid,
  include_data = FALSE,
  is_submission = FALSE,
  build_dir = tempdir(),
  params = list(),
  report_format = c("html", "pdf")
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report
instance.}

\item{zipfile}{\code{character(1)}. The zip file to which the contents of
the report preview will be saved.}

\item{include_image}{\code{logical(1)}. When \code{TRUE}, any images that are
created as part of the report will be included in the supplemental
files with the preview.}

\item{include_data}{\code{logical(1)}. When \code{TRUE}, supporting data generated
by the report will be included with the supplemental files. Note: this
is unrelated to the option to include data with report submissions that
is part of the report template.}

\item{build_dir}{\code{character(1)}. The directory into which the report
files will be saved.}

\item{params}{\code{list}. Optional parameters to send to the report template.
Note: the report intance OID, report template OID, and \code{build_dir} are
included already.}

\item{report_format}{\code{character(1)}. One of \code{c("html", "pdf")}. The
format of the report being generated.}

\item{is_submission}{\code{logical(1)}. When \code{TRUE}, the report is generated
as a submission (with a revision number).}

\item{report_filename}{\code{character(1)}. full path to the file created
by \code{rmarkdown::build}. Typically this is used to locate supplemental
files, such as images and supporting data.}

\item{files_to_zip}{\code{character}. The full path to the files to be included
in the zip file. Include the report file, images, and supplemental data.}

\item{is_preview}{\code{logical(1)}. When \code{TRUE}, the report is generated as
a preview. Otherwise it is generated ready for archival.}
}
\description{
Generates the report, supplemental data, and image files
that are included in the zip files or attached to an e-mail.
}
