% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInstanceToAutoDistribute.R
\name{getInstanceToAutoDistribute}
\alias{getInstanceToAutoDistribute}
\title{Determine if Report Instances are Scheduled for Automatic Distribution}
\usage{
getInstanceToAutoDistribute(distribution_run_time = Sys.time())
}
\arguments{
\item{distribution_run_time}{\code{POSIXct(1)} The time at which the
distribution is being run.}
}
\value{
Returns a data frame with the columns:

\tabular{ll}{
\code{OID} \tab The OID of the ReportTemplateAutoDistribute object. \cr
\code{ParentReportTemplateOID} \tab The OID of the ReportTemplate object. \cr
\code{TemplateName} \tab The name of the Report Template. \cr
\code{IsAddToArchive} \tab \code{logical} indicating if the report is added to the archive when distributed. \cr
\code{ReportFormat} \tab Either "html" or "pdf". The delivery format for the report. \cr
\code{AutoDistributeDate} \tab The most recent qualifying time for the report's distribution. \cr
\code{ReportInstanceOID} \tab The OID of the ReportInstance object.
}
}
\description{
Returns a data frame with the instances that are scheduled
for automatic distribution but have not yet been distributed.
}
