% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeReportManagerDatabase.R
\name{initializeReportManagerDatabase}
\alias{initializeReportManagerDatabase}
\title{Intialize the Report Manager Database}
\usage{
initializeReportManagerDatabase(filename)
}
\arguments{
\item{filename}{\code{character(1)}. A filename of SQL code. Usually
one found in \code{system.file("Sql", package = "ReportManager")}}
}
\description{
Retrieves the SQL code to define the Report Manager
database tables and runs it through the appropriate connection.
}
