% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inputValidationCollection.R
\name{inputValidationCollection}
\alias{inputValidationCollection}
\title{Create an Input Validation Collection}
\usage{
inputValidationCollection()
}
\description{
Creates an objet to assist in collecting input validation
findings and reporting them to the user.
}
