% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertRecord.R
\name{insertRecord}
\alias{insertRecord}
\alias{updateRecord}
\title{Insert and Update Records}
\usage{
insertRecord(
  data,
  table_name,
  return_oid = TRUE,
  flavor = getOption("RM_sql_flavor"),
  id_field = "OID",
  schema = "dbo"
)

updateRecord(
  data,
  where_data,
  table_name,
  flavor = getOption("RM_sql_flavor"),
  schema = "dbo"
)
}
\description{
Generalized forms for writing and executing statements
to add new records or update existing records.
}
