% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_application.R
\name{launch_application}
\alias{launch_application}
\title{Launch the Reports Manager Application}
\usage{
launch_application(app_dir = character(0), ...)
}
\arguments{
\item{app_dir}{\code{character} with a maximum length of 1. Path to the 
directory with the Reports Manager application}

\item{...}{Additional arguments to pass to \code{\link[shiny]{runApp}}}
}
\description{
Starts the Reports Manager Shiny Application
}
\author{
Benjamin Nutter
}
