% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_database_connection.R
\name{make_database_connection}
\alias{make_database_connection}
\alias{close_database_connection}
\title{Make a Database Connection}
\usage{
make_database_connection(db_path = getOption("report_manager_db_path"))

close_database_connection(conn)
}
\arguments{
\item{db_path}{\code{character(1)} For demonstration, the path 
to the MS Access database with the Report Manager database.}

\item{conn}{Database connection to be closed.}
}
\description{
The reporting application is heavily dependent on 
  reading from and writing to its associated database. This
  wrapper function is intended to manage those connections.
}
\details{
Because the Reports Manager is designed to be adaptable
  to any database system, this function is one that must be 
  modified and customized to the unique environment where it
  is being implemented. In the development version on GitHub, 
  it connects to the MS Access Database included in the package
  files.
}
\author{
Benjamin Nutter
}
