% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReportFileName.R
\name{makeReportFileName}
\alias{makeReportFileName}
\title{Make File Names for Report Instances}
\usage{
makeReportFileName(
  report_instance_oid,
  is_preview = TRUE,
  is_submission = FALSE,
  file_extension = c("zip", "html", "pdf")
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The start time of the reporting period.
This will be truncated to just the date portion in the filename.}

\item{is_preview}{\code{logical(1)}. When \code{TRUE}, the "Preview" modifier is
added to the file name.}

\item{is_submission}{\code{logical(1)}. Then \code{TRUE}, the next revision number
is added to the file name.}

\item{file_extension}{\code{character(1)}. One of \code{c("zip", "html", "pdf")}.}
}
\description{
Produces a standardized file name for report instances
when downloading a preview or generating a report.
}
\details{
File names take the format of
'\link{start_date} - \link{end_date}-\link{is_preview}-\link{template_name}-Rev\link{number}.\link{file_extension}'.
}
