% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeSignatureTable.R
\name{makeSignatureTable}
\alias{makeSignatureTable}
\title{Make the UI Table for Signing Reports}
\usage{
makeSignatureTable(signature_data)
}
\arguments{
\item{signature_data}{\code{data.frame}. Must have the columns
\code{c("ParentRole", "RoleName", "IsSigned", "SignatureDateTime", "IsSigned")}.}
}
\description{
Uses the Signature data from ReportInstanceSignature
to produce a table with user controls for signing reports and
removing the signatures.
}
