% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryAutoDistribution.R
\name{queryAutoDistribution}
\alias{queryAutoDistribution}
\title{Retrieve Auto Distribution Configuration from the Database}
\usage{
queryAutoDistribution(oid = numeric(0), report_template_oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the AutoDistribution object.}

\item{report_template_oid}{\code{integerish(0/1)}. The OID of the report template.
If none is provided, all records are returned.}
}
\description{
Provides the user with tools to retrieve the configuration
settings of automated report distribution.
}
