% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryCurrentRevisionNumber.R
\name{queryCurrentRevisionNumber}
\alias{queryCurrentRevisionNumber}
\title{Retrieve Current Revision Number for Active Report Submission}
\usage{
queryCurrentRevisionNumber(report_instance_oid)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance.}
}
\description{
Returns the number of the current revision in progress.
If no submissions have been made, it starts at 0.
}
