% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDateReportingFormat.R
\name{queryDateReportingFormat}
\alias{queryDateReportingFormat}
\title{Retrieve DateReportingFormat Objects from the Database}
\usage{
queryDateReportingFormat(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the DataReportingFormat object
to return. When length 0, all objects are returned.}
}
\description{
Executes a query to retrieve DataReportingFormat objects
from the database.
}
