% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryDisclaimer.R
\name{queryDisclaimer}
\alias{queryDisclaimer}
\title{Retrieve Disclaimer Objects from the Database}
\usage{
queryDisclaimer(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the Disclaimer object to retrieve.
When length is 0, all objects are returned.}
}
\description{
Retrieves Disclaimer objects from the database.
}
