% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFileArchive.R
\name{queryFileArchive}
\alias{queryFileArchive}
\title{Retrieve FileARchive Records}
\usage{
queryFileArchive(
  oid = numeric(0),
  parent_report_template = numeric(0),
  parent_report_instance = numeric(0)
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the FileArchive object to retrieve.}

\item{parent_report_template}{\code{integerish(0/1)}. The OID of the
ReportTemplate object with which the file is associated.}

\item{parent_report_instance}{\code{integerish(0/1)}. The OID of the
ReportInstance object with which the file is associated.}
}
\description{
Provides the user an interface to retrieve FileArchive
objects from the database.
}
