% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFooter.R
\name{queryFooter}
\alias{queryFooter}
\title{Retrieve Footer Objects from the Database}
\usage{
queryFooter(oid)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the Footer object to retrieve.
When length is 0, all objects are returned.}
}
\description{
Retrieves Footer objects from the database.
}
