% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryFromFileArchive.R
\name{queryFromFileArchive}
\alias{queryFromFileArchive}
\title{Get a File from the FileArchive}
\usage{
queryFromFileArchive(oid, file_dir = character(0))
}
\arguments{
\item{oid}{\code{integerish(1)}. The OID of the file to retrieve from the
FileArchive.}

\item{file_dir}{\code{character(0/1)}. A directory to which the file is to
be saved. If the length is zero, the save is skipped and only the
data frame with the file BLOB object is returned.}
}
\description{
This function enables the user to retrieve a file from the
FileArchive table.
}
\seealso{
\code{\link[=queryFileArchive]{queryFileArchive()}}, \cr
\code{\link[=queryLogo]{queryLogo()}}
}
