% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryInstanceDistributionSelection.R
\name{queryInstanceDistributionSelection}
\alias{queryInstanceDistributionSelection}
\title{Retrieve the Distribution List for a Report Instance}
\usage{
queryInstanceDistributionSelection(report_instance_oid)
}
\arguments{
\item{report_instance_oid}{\verb{integerish(1).} The OID of the report instance.}
}
\description{
While the distribution may be set at a template level,
each instance of the report template may modify that distribution list.
This function executes a query to combine the template-level and
instance-level distribution list.
}
