% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryLogo.R
\name{queryLogo}
\alias{queryLogo}
\title{Retrieve FileArchive Objects Marked as Logos}
\usage{
queryLogo()
}
\description{
Retrieves the FileArchive objects marked as Logos. These
can be made available to display with the title of the report.
}
