% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportInstance.R
\name{queryReportInstance}
\alias{queryReportInstance}
\title{Retrieve Report Instances from the Database}
\usage{
queryReportInstance(
  report_template_oid = numeric(0),
  report_instance_oid = numeric(0)
)
}
\arguments{
\item{report_template_oid}{\code{integerish(0/1)}. The OID of the report
template for which instances are returned. If length 0, instances for
all report templates are returned.}

\item{report_instance_oid}{\code{integerish(0/1)}. The OID of the report
instance to be queried. If length 0, all instances are returned.
Note: if both \code{report_template_oid} and \code{report_instance_oid} are
provided and the instance OID does not belong to the template OID,
no results will be returned.}
}
\value{
Returns a data frame of the instance details for a report template.
}
\description{
Executes a query to the database and retrieves the report
instances associated with the report template provided. May also
return the instance data for a specific instance.
}
