% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportInstanceNarrative.R
\name{queryReportInstanceNarrative}
\alias{queryReportInstanceNarrative}
\title{Retrieve the Narrative for a Report Instance}
\usage{
queryReportInstanceNarrative(report_instance_oid)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance
for which the narrative is to be retrieved.}
}
\description{
Enables the user to query the database for the current
associated with a report instance.
}
