% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportInstanceNarrativeEvent.R
\name{queryReportInstanceNarrativeEvent}
\alias{queryReportInstanceNarrativeEvent}
\title{Retrieve the History of Narrative Editing Events for a Report Instance}
\usage{
queryReportInstanceNarrativeEvent(report_instance_oid)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance
for which the narrative event history is to be retrieved.}
}
\description{
Enables the user to query the database for the history of
events where the narrative associated with a report instance was edited.
}
