% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportInstanceNote.R
\name{queryReportInstanceNote}
\alias{queryReportInstanceNote}
\title{Query Database for Report Instance Notes}
\usage{
queryReportInstanceNote(report_instance_oid)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance.}
}
\description{
Queries the database for notes associated with a
report instance.
}
