% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportSelection.R
\name{queryReportSelection}
\alias{queryReportSelection}
\title{Retrieve Report Templates Configured for Reporting}
\usage{
queryReportSelection()
}
\value{
Retuns a \code{data.frame} with attributes of the report template needed for
navigating the generation of reports in the UI.
}
\description{
Retrieve from the database the report templates that have
been configured for reporting along with information necessary to
generate previews.
}
