% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportTemplate.R
\name{queryReportTemplate}
\alias{queryReportTemplate}
\title{Retrieve Report Template Settings}
\usage{
queryReportTemplate(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReportTemplate object for
which data is to be retrieved. Use \code{numeric(0)} to return all templates.}
}
\description{
Provides the user with an interface to retrieve configuration
settings of ReportTemplate objects. This will only return the settings
in the primary ReportTemplate object. Other functions will give access
to settings that are related through other tables.
}
