% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportDistribution.R,
%   R/queryReportTemplateDistribution.R
\name{queryReportTemplateDistribution}
\alias{queryReportTemplateDistribution}
\title{Retrieve ReportTemplateDistribution Objects}
\usage{
queryReportTemplateDistribution(
  oid = numeric(0),
  parent_report_template = numeric(0),
  parent_user = numeric(0),
  parent_role = numeric(0)
)

queryReportTemplateDistribution(
  oid = numeric(0),
  parent_report_template = numeric(0),
  parent_user = numeric(0),
  parent_role = numeric(0)
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. When length is 1, the OID of the
ReportTemplateSignature to be retrieved. Otherwise, all
records are returned.}

\item{parent_report_template}{\code{integerish(0/1)}. When length is 1, the
OID of the ReportTemplate to include in the query clause. Otherwise,
all records are returned.}

\item{parent_user}{\code{integerish(0/1)}. When length is 1, the OID
of the User to include in the query clause. Otherwise, all
records are returned.}

\item{parent_role}{\code{integerish(0/1)}. When length is 1, the OID
of the Role to include in the query clause. Otherwise, all
records are returned.}
}
\description{
Provides the user with an interface to retrieve
ReportTemplateDistribution objects.

Provides the user with an interface to retrieve
ReportTemplateDistribution objects.
}
\details{
The WHERE clause for the query is constructed using any of the
\code{oid}, \code{parent_report_template}, or \code{parent_user} arguments
that have a non-zero length. The query returns all records that
satisfy all of the conditions.

The WHERE clause for the query is constructed using any of the
\code{oid}, \code{parent_report_template}, \code{parent_user} or \code{parent_role} arguments
that have a non-zero length. The query returns all records that
satisfy all of the conditions.  The ReportTemplateDistribution object
should only have either a ParentUser or ParentRole value. If both
\code{parent_user} and \code{parent_role} are given, it is expected that no
rows will be returned.
}
