% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportTemplateFooter.R
\name{queryReportTemplateFooter}
\alias{queryReportTemplateFooter}
\title{Retrieve ReportTemplateFooter Objects}
\usage{
queryReportTemplateFooter(
  oid = numeric(0),
  parent_report_template = numeric(0),
  parent_footer = numeric(0)
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. When length is 1, the OID of the
ReportTemplateFooter to be retrieved. Otherwise, all
records are returned.}

\item{parent_report_template}{\code{integerish(0/1)}. When length is 1, the
OID of the ReportTemplate to include in the query clause. Otherwise,
all records are returned.}

\item{parent_footer}{\code{integerish(0/1)}. When length is 1, the OID
of the Disclaimer to include in the query clause. Otherwise, all
records are returned.}
}
\description{
Provides the user with an interface to retrieve
ReportTemplateFooter objects.
}
\details{
The WHERE clause for the query is constructed using any of the
\code{oid}, \code{parent_report_template}, or \code{parent_disclaimer} arguments
that have a non-zero length. The query returns all records that
satisfy all of the conditions.
}
