% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportSignature.R
\name{queryReportTemplateSignature}
\alias{queryReportTemplateSignature}
\title{Retrieve ReportTemplateSignature Objects}
\usage{
queryReportTemplateSignature(
  oid = numeric(0),
  parent_report_template = numeric(0),
  parent_role = numeric(0)
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. When length is 1, the OID of the
ReportTemplateSignature to be retrieved. Otherwise, all
records are returned.}

\item{parent_report_template}{\code{integerish(0/1)}. When length is 1, the
OID of the ReportTemplate to include in the query clause. Otherwise,
all records are returned.}

\item{parent_role}{\code{integerish(0/1)}. When length is 1, the OID
of the Role to include in the query clause. Otherwise, all
records are returned.}
}
\description{
Provides the user with an interface to retrieve
ReportTemplateSignature objects.
}
\details{
The WHERE clause for the query is constructed using any of the
\code{oid}, \code{parent_report_template}, or \code{parent_role} arguments
that have a non-zero length. The query returns all records that
satisfy all of the conditions.
}
