% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportTemplateUserPermission.R
\name{queryReportTemplateUserPermission}
\alias{queryReportTemplateUserPermission}
\title{Retrieve the Permissions Granted a User on a Report Template}
\usage{
queryReportTemplateUserPermission(parent_report_template, parent_user)
}
\arguments{
\item{parent_report_template}{\code{integerish(1)} The OID of the Report Template
for which permissions are being checked.}

\item{parent_user}{\code{integerish(1)}. The OID of the User for whom permissions
are being checked.}
}
\description{
Enables the user to retrieve the permissions granted to a user
on a selected report template. Typically this is used to determine if
the current user has permission to perform actions on a template.
}
