% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryReportUser.R
\name{queryReportUser}
\alias{queryReportUser}
\title{Fetch the ReportUser Table}
\usage{
queryReportUser(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the Report User to fetch
from the database. By default (\code{character(0)}), all records are
returned.}
}
\description{
Fetch the ReportUser table from the database.
}
