% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryRole.R
\name{queryRole}
\alias{queryRole}
\title{Query Roles From the ReportManager Database}
\usage{
queryRole(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. When length is 0, all roles are returned.
Otherwise, only the requested role is returned.}
}
\description{
Retrieves the roles from the database.
}
