% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/querySchedule.R
\name{querySchedule}
\alias{querySchedule}
\title{Retrieve Schedule Objects from the Database}
\usage{
querySchedule(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. When length 0, all schedule objects are
returned. When length 1, only the schedule matching \code{oid} is returned.}
}
\description{
Executes a query to get the Schedule objects from the
database.
}
