% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryUser.R
\name{queryUser}
\alias{queryUser}
\title{Fetch the User Table}
\usage{
queryUser(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the Report User to fetch
from the database. By default (\code{character(0)}), all records are
returned.}
}
\description{
Fetch the User table from the database.
}
