% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryUserRole.R
\name{queryUserRole}
\alias{queryUserRole}
\title{Retrieve User Role Assignments}
\usage{
queryUserRole(oid = numeric(0))
}
\arguments{
\item{oid}{\code{integerish(0/1)}. When length 0, all records are returned.
Otherwise, only the requested record is returned.}
}
\description{
Executes a query to retrieve the UserRole assignments from the
database.
}
