% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/queryUserSignature.R
\name{queryUserSignature}
\alias{queryUserSignature}
\title{Retrieve User Signature Files}
\usage{
queryUserSignature(user_oid = numeric(0))
}
\arguments{
\item{user_oid}{\code{integerish(0/1)}. The OID of the user. If the length is zero,
all records are returned.}
}
\description{
Returns one or more User Signature Files from the database.
}
