% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioDataTable.R
\name{radioDataTable}
\alias{radioDataTable}
\title{Make a Data Table Element with Radio Buttons in the First Column}
\usage{
radioDataTable(data, id_variable, element_name, checked = character(0))
}
\arguments{
\item{data}{\code{data.frame} table of data.}

\item{id_variable}{\code{character(1)}. The name of a column in \code{data}
that will be used to make the selection IDs.}

\item{element_name}{\code{character(1)}. The name for the page element. This
acts as the shiny \code{inputId}.}

\item{checked}{\code{integerish(0/1)} or \code{character(0/1)}. The value
from \code{id_variable} that should be marked checked.}
}
\description{
Constructs a Data Table element that can be used with
radio buttons to make selections.
}
\author{
Benjamin Nutter
}
