% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/radioDataTableOutput.R
\name{radioDataTableOutput}
\alias{radioDataTableOutput}
\title{Display a DataTable Object with Radio Buttons}
\usage{
radioDataTableOutput(
  outputId,
  radioId,
  height = "1000px",
  width = "100\%",
  horizontal_scroll = TRUE
)
}
\arguments{
\item{outputId}{\code{character(1)} The name of the output created using
\code{\link[DT:dataTableOutput]{DT::renderDataTable()}}.}

\item{radioId}{\code{character(1)} Should match the \code{element_name}
argument of the call to \code{makeRadioDataTable}}

\item{height}{\code{character(1)} The height of the element.}

\item{width}{\code{character(1)} The width of the element.}

\item{horizontal_scroll}{\code{logical(1)} Allow horizontal scrolling for wide tables?}
}
\description{
Creates the User Interface objects for a DataTable object
that includes radio buttons. This is used in the ui.R object to
opposite server output objects that user \code{\link[DT:dataTableOutput]{DT::renderDataTable()}}
with a call to \code{makeRadioDataTable}.
}
\author{
Benjamin Nutter
}
