% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sendEmail.R
\name{sendEmail}
\alias{sendEmail}
\title{Send an Email from the Application}
\usage{
sendEmail(
  from_user_oid,
  to_address,
  message,
  report_template,
  report_instance_oid,
  filename = character(0),
  control = list(smtpServer = getOption("RM_smtpServer")),
  embed_html = FALSE,
  is_revision = FALSE,
  ...
)
}
\arguments{
\item{from_user_oid}{\code{intgerish(1)}. The OID of the user initiating the
send action.}

\item{to_address}{\code{character(n)}. The email address to which the e-mail
will be sent.}

\item{message}{\code{character(1)}. The message included in the body of the e-mail.}

\item{report_template}{\code{data.frame} of the Report Template settings.}

\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance
being sent.}

\item{filename}{\code{character(n)}. The files to attach to the e-mail.}

\item{control}{\code{list} control parameters to pass to the argument of the
same name in \code{\link[sendmailR]{sendmail}}.}

\item{embed_html}{\code{logical(1)}. Controls if the content is embedded into the
email or attached as a file.}

\item{is_revision}{\code{logical(1)}. Indicates if the e-mail is being sent
regarding a revision.}

\item{...}{Other arguments to pass to \code{\link[sendmailR]{sendmail}}.}
}
\description{
Sends and e-mail from the Report Manager Application
}
