% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startReportManager.R
\name{startReportManager}
\alias{startReportManager}
\title{Start the Report Manager Application}
\usage{
startReportManager(email, ...)
}
\arguments{
\item{email}{\code{character(1)}. May be either \code{"live"}, \code{"off"} or an
email address. When \code{"live"}, emails generated by the application are
sent in accordance with the configurations in the database. When
\code{"off"}, no e-mails are sent. If the user provides an e-mail address,
all e-mails will be sent to the address provided.}

\item{...}{Arguments to pass to \code{\link[shiny:runApp]{shiny::runApp()}}, such as
\code{launch.browser} or \code{port}.}
}
\description{
Starts the Report Manager Application
}
