% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submitReport.R
\name{submitReport}
\alias{submitReport}
\title{Submit a Report Instance}
\usage{
submitReport(
  report_instance_oid,
  is_submission,
  is_distribute,
  is_distribute_internal_only,
  is_add_to_archive,
  is_embed_html,
  message = "",
  params = list(),
  report_format,
  current_user_oid
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance.}

\item{is_submission}{\code{logical(1)}. When \code{TRUE}, the report is being counted
as a submission and will be assigned a revision number.}

\item{is_distribute}{\code{logical(1)}. When \code{TRUE}, the report will be distributed
via e-mail.}

\item{is_distribute_internal_only}{\code{logical(1)}. When \code{TRUE}
(and \code{is_distribute} is \code{TRUE}), the report is only distributed to internal
recipients. (Ignored when \code{is_distribute} is \code{FALSE}).}

\item{is_add_to_archive}{\code{logical(1)}. When \code{TRUE}, report files are added
to the file archive.}

\item{is_embed_html}{\code{logical(1)}. When \code{TRUE} (and the \code{report_format} is
\code{"html"}), the report content is embedded in the e-mail (instead of
attached).}

\item{params}{named \code{list}. Parameters to pass to the report template.}

\item{report_format}{\code{character(1)}. The file format for the report. One of
\code{c("html", "pdf")}.}

\item{current_user_oid}{\code{integerish(1)}. The OID of the current user OID.}
}
\description{
Generate report files, add them to the archive, and
submit them to internal/external clients.
}
