% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateInstanceIsSubmitted.R
\name{updateInstanceIsSubmitted}
\alias{updateInstanceIsSubmitted}
\title{Update IsSubmitted Property on a Report Instance}
\usage{
updateInstanceIsSubmitted(
  report_instance_oid,
  is_submitted,
  current_user_oid,
  event_date_time = Sys.time()
)
}
\arguments{
\item{report_instance_oid}{\code{integerish(1)}. The OID of the report instance}

\item{is_submitted}{\code{logical(1)}. The value to which the IsSubmitted
property will be set.}

\item{current_user_oid}{\code{integerish(1)}. The OID of the user performing
the action.}

\item{event_date_time}{\code{POSIXct(1)}. The date/time at which the action
was performed}
}
\description{
Changes the IsSubmitted property on a Report Instance record.
}
