% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateReportInstanceSchedule.R
\name{updateReportInstanceSchedule}
\alias{updateReportInstanceSchedule}
\title{Update the Report Instance Schedule of a Report Template}
\usage{
updateReportInstanceSchedule(report_template_oid, event_user)
}
\arguments{
\item{report_template_oid}{\code{integerish(1)}. The OID of the report template
for which the schedule should be updated.}

\item{event_user}{\code{integerish(1)}. the User.OID of the user performing
the action.}
}
\description{
Retrieves the schedule information of a report template,
determines the report schedule, and adds any unrecorded instances
to the database.
}
