% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditReportTemplate.R
\name{addEditReportTemplate}
\alias{addEditReportTemplate}
\title{Add or Edit ReportTemplate Objects}
\usage{
addEditReportTemplate(
  oid = numeric(0),
  template_directory,
  template_file,
  title,
  title_size,
  include_toc = FALSE,
  default_email = "",
  is_signature_required = FALSE,
  is_active = TRUE,
  logo_oid = numeric(0),
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the ReportTemplate object to be
edited. Use \code{numeric(0)} to add a new object.}

\item{template_directory}{\code{character(1)}. The directory in which the
report template files are stored. No more than 50 characters.}

\item{template_file}{\code{character(1)}. The primary template file that
directs the progress of the report template. No more than 50 characters.}

\item{title}{\code{character(1)}. The title to display on the report. No more
than 200 characters.}

\item{title_size}{\code{character(1)}. The size of the font to use when
rendering the title.}

\item{include_toc}{\code{logical(1)}. When \code{TRUE}, a table of contents will be
included in the report.}

\item{default_email}{\code{character(1)}. The default text to display in the
e-mail submission dialog box.}

\item{is_signature_required}{\code{logical(1)}. When \code{TRUE}, the template
requires signatures before it can be delivered to external users.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the template is flagged for
active use.}

\item{logo_oid}{\code{integerish(0/1)}. The OID of the logo to display when
rendering the title.}

\item{event_user}{\code{integerish(1)}. The OID of the user performing the
action.}
}
\description{
Enables the user to add ReportTemplate objects or edit
existing objects. Only the properties stored directly on
the ReportTemplate object are edited with this function.
}
