% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addEditSchedule.R
\name{addEditSchedule}
\alias{addEditSchedule}
\title{Add or Edit a Schedule Object}
\usage{
addEditSchedule(
  oid = numeric(0),
  schedule_name,
  frequency,
  frequency_unit,
  offset_overlap = 0,
  offset_overlap_unit = "None",
  is_active = TRUE,
  is_period_to_date = FALSE,
  event_user
)
}
\arguments{
\item{oid}{\code{integerish(0/1)}. The OID of the schedule to be edited.
When the length is 0, a new schedule will be added.}

\item{schedule_name}{\code{character(1)}. The name of the schedule.}

\item{frequency}{\code{integerish(1)}. The frequency at which the schedule repeats.}

\item{frequency_unit}{\code{character(1)}. The units associated with the frequency.
One of \code{c("Minute", "Hour", "Day", "Week", "Month", "Year")}.}

\item{offset_overlap}{\code{integerish(1)}. The duration of scheduling overlap.
This allows for schedules with overlapping reporting periods.}

\item{offset_overlap_unit}{\code{character(1)}. The units associated with the
offset overlap. One of \code{c("None", "Minute", "Hour", "Day", "Week", "Month", "Year")}.}

\item{is_active}{\code{logical(1)}. When \code{TRUE}, the schedule is marked as
active.}

\item{is_period_to_date}{\code{logical(1)}. When \code{TRUE}, the schedule will take
the form of a "period-to-date" (year-to-date, for example). These schedules
may have multiple instances with the same start date but end dates that
progress throughout the periodicity.}

\item{event_user}{\code{integerish(1)}. the User.OID of the user performing
the action.}
}
\description{
Adds a new schedule object or edits an existing schedule
objects in the database.
}
