% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectToReportManager.R
\name{connectToReportManager}
\alias{connectToReportManager}
\title{Connect to the Report Manager Database}
\usage{
connectToReportManager(flavor = getOption("RM_sql_flavor"), ...)
}
\arguments{
\item{flavor}{\code{character(1)}. The flavor of SQL used. This will
determine the SQL engine used. One of \code{c("sql_server", "sqlite")}.}

\item{...}{Arguments for establishing the connection via the \code{DBI} package.}
}
\description{
Establish a Connection to the Report Manager Database.
}
