% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initializeReportManagerDatabase.R
\name{initializeReportManagerDatabase}
\alias{initializeReportManagerDatabase}
\alias{initializeUiTestingDatabase}
\title{Intialize the Report Manager Database}
\usage{
initializeReportManagerDatabase(
  filename,
  last_name,
  first_name,
  login_id,
  email
)

initializeUiTestingDatabase(
  filename,
  include = c("User", "Role", "UserRole"),
  last_name = "Testing-LastName",
  first_name = "Testing-FirstName",
  login_id = Sys.info()["user"],
  email = "email@somewhere.net"
)
}
\arguments{
\item{filename}{\code{character(1)}. A filename of SQL code. Usually
one found in \code{system.file("Sql", package = "ReportManager")}}

\item{last_name}{\code{character(1)}. The last name of the user.}

\item{first_name}{\code{character(1)}. The first name of the user.}

\item{login_id}{\code{character(1)}. The login ID of the user.}

\item{email}{\code{character(0/1)}. The e-mail address of the user. If not
provided, the user will not receive reports distributed by e-mail but
may use the application otherwise.}
}
\description{
Retrieves the SQL code to define the Report Manager
database tables and runs it through the appropriate connection.
}
\details{
\code{last_name}, \code{first_name}, \code{login_id}, and \code{email} are
the values assigned for the first user in the database. This user
will be given the UserAdministrator and ReportAdministrator roles,
allowing them to add other users and configure reports.
}
